+++ eth-contracts/contracts/test/StakingUpgraded.sol
@@ -4,7 +4,7 @@ import "../Staking.sol";
 
 
 contract StakingUpgraded is Staking {
-    function newFunction() public view returns (uint) {
+    function newFunction() public view returns (uint256) {
         _requireIsInitialized();
 
         return 5;
